requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.30.6
jinja2==3.1.4
python-multipart==0.0.9
sqlalchemy==2.0.36
asyncpg==0.29.0
alembic==1.13.3
pydantic==2.10.2
pydantic-settings==2.6.1
passlib[bcrypt]==1.7.4
itsdangerous==2.2.0
jsonschema==4.23.0
apscheduler==3.10.4
httpx==0.27.2
python-dateutil==2.9.0.post0
